+++ NEWS.md
@@ -1,3 +1,7 @@
+## rlistings 0.2.10
+ * Added an error message for listings with variables of `difftime` class.
+ * Added message when the listing object has zero row.
+
 ## rlistings 0.2.9
  * Added `truetype` font support based on new `formatters` api, by @gmbecker.
  * Fixed tests so that paginations based on different fonts and page sizes can be compared, by @gmbecker.